feat : 채팅 연결 완료

+++ android/Signal/app/src/main/java/com/ongo/signal/data/model/chat/ChatHomeCreateDTO.kt
@@ -3,6 +3,6 @@ package com.ongo.signal.data.model.chat
 import com.google.gson.annotations.SerializedName
 
 data class ChatHomeCreateDTO(
-    @SerializedName("from_id") val fromId : Int,
-    @SerializedName("to_id")val toId : Int,
+    @SerializedName("from_id") val fromId : Long,
+    @SerializedName("to_id")val toId : Long,
 )
